refactor: :truck: RoomDetail 페이지 이동
