--- src/features/room/components/RoomDetail/RoomDetail.types.ts
@@ -1,1 +1,0 @@
-export interface RoomDetailProps {}
