--- src/features/room/components/RoomDetail/index.ts
@@ -1,2 +1,0 @@
-export { default } from './RoomDetail';
-export type { RoomDetailProps } from './RoomDetail.types';
