+++ src/features/room/components/index.ts
@@ -1,5 +1,4 @@
 export { default as RoomMemoForm } from './RoomMemoForm';
-export { default as RoomDetail } from './RoomDetail';
 export { default as RoomDetailMenu } from './RoomDetailMenu';
 export { default as RoomTypeRadioGroup } from './RoomTypeRadioGroup';
 export { default as UpsertRoomDialog } from './UpsertRoomDialog';
