+++ src/pages/rooms/[id]/detail.styles.ts
