+++ src/pages/rooms/[id]/detail.types.ts
@@ -1,0 +1,3 @@
+export interface RoomDetailProps {
+  id: string;
+}
